+++ src/main/java/lotto/Application.java
@@ -1,7 +1,12 @@
 package lotto;
 
+import lotto.controller.LottoController;
+import lotto.model.LottoModel;
+import lotto.view.LottoView;
+
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+        LottoController controller = new LottoController(new LottoModel(), new LottoView());
+        controller.run();
     }
 }
